requirements.txt
torch==1.3.1
numpy==1.17.2
matplotlib==3.1.1
Pillow==6.1.0
